+++ marshmallow/__init__.py
@@ -13,9 +13,8 @@ from marshmallow.decorators import (
 from marshmallow.utils import pprint, missing
 from marshmallow.exceptions import ValidationError
 
-__version__ = '2.2.1'
+__version__ = '2.3.0'
 __author__ = 'Steven Loria'
-__license__ = 'MIT'
 
 __all__ = [
     'Schema',