+++ marshmallow/marshalling.py
@@ -131,10 +131,9 @@ class Marshaller(ErrorStore):
         for attr_name, field_obj in iteritems(fields_dict):
             if getattr(field_obj, 'load_only', False):
                 continue
-            if not self.prefix:
-                key = attr_name
-            else:
-                key = ''.join([self.prefix, attr_name])
+
+            key = ''.join([self.prefix or '', field_obj.dump_to or attr_name])
+
             getter = lambda d: field_obj.serialize(attr_name, d, accessor=accessor)
             value = self.call_and_store(
                 getter_func=getter,